fixed failed tests, separated ProgressContext from progress.h
